OKAY  I SWEAR THIS ONE WORKS THIS TIME
I have made some changes regarding the names of some files but they are temporary for now. As we progress we will find better names for them.

The only thing that was really necessary was deleting any files that made each person's code be recognized as an individual app, and installing some extra dependent react libraries.