I have no idea what is wrong
For some reason, whenever I try to pull from the database, no matter what I set .ref to, it ALWAYS accesses the same node. It only gets data from the node "listings" even if I specify a different one. Now, despite "listings" not being in the database, it will still return the data it once had. So, if the node is no longer in the database, whatever its previous state was, the phone will ALWAYS load it. It makes zero sense.