Gets images from database
images are retrieved from database by getting a url of the image that is stored in the database. however, when someone create a new listing we will have to push the image to storage, get its url, and then save it in the databse. SO maybe my imeplementation will change in the future but it's currently like this so okay bye <3