button group and text input
added a rudimentary button group and text input for the form. backgrounds are styled different colors so I can see them. I also moved the customComponents folder into the root of the project so we can keep track of everyone's custom components. I have made edits to imports in various files to account for the relocation of the customComponents directory