Wbeeson wf create listing (#6)
* test commit

testing commiting and trying to merge with master

* fixed issue with custom fonts

Continued to get error "fontfamily MaterialIcons is not a system font". I bet you guys will encounter it later too. Check the wiki for documentation about how to fix it in future.

* Rudimentary search function

currently compares search text with WHOLE descriptions and titles. eventually will need to tokenize titles and descriptions and then compare the input with the tokens.

* added firebase library

* Added gitignore and some database accessing to get listings cells

does not incldue FIre.js firebase initialization. store somewhere else in a top secret location offshore

* I have no idea what is wrong

For some reason, whenever I try to pull from the database, no matter what I set .ref to, it ALWAYS accesses the same node. It only gets data from the node "listings" even if I specify a different one. Now, despite "listings" not being in the database, it will still return the data it once had. So, if the node is no longer in the database, whatever its previous state was, the phone will ALWAYS load it. It makes zero sense.

* no more folder

removed unecessary folder

* fix

yes

* Gets images from database

images are retrieved from database by getting a url of the image that is stored in the database. however, when someone create a new listing we will have to push the image to storage, get its url, and then save it in the databse. SO maybe my imeplementation will change in the future but it's currently like this so okay bye <3

* Added profile thing op top right

icon on top right that WILL lead to the profile page. currently does not.

* added icons to tab bar

i just added icons from expo's vector icons. nothing else functional

* I don't know what I did

here ya go

* pull to refresh

added the ability to pull the list down in order to refresh from database.

* button group and text input

added a rudimentary button group and text input for the form. backgrounds are styled different colors so I can see them. I also moved the customComponents folder into the root of the project so we can keep track of everyone's custom components. I have made edits to imports in various files to account for the relocation of the customComponents directory

* very not pleasing

looks like garbage, but is the basic layout. i'm working on styling. so in other words my rate of expiration is increasing exponentially

* more consistant styling

more consistant looking but need to make sure the sizes of the cards are a good size

* okay i swear this time the styling is good

actually decent styling

* empty listing images section

an empty card that will eventually allow users to include their own images. will require native code and work within XCode and Android Studio.

* made some edits, added a new button

new button, added some warning messages, edited the statusbar padding style

* submit button response to form

submit button is disabled when the required informaiton is not filled out. will enable when all required information is filled out

* field required messages appear and disappear

depending on if info is entered, the error messages will show or hide

* changed gitignore

* Delete Fire.js

* pushes node to database

based on the listing form, pushes a node to the database with all the form data.

* implementing tab and stack navigators

it kinda works.... not really though. still a lot of work to do for it

* hiding stuff

* everything is in order

nav elements hide appropriately, status bar padding, listing creation page works and submits correctly and navigates correctly, refresh works, everything so far LOOKS like it works.

* this is the new thing we are doing thank you

hello it is we

* sure why not

just some stuff git is making me do in order to merge into master